Add Base classes for FormCells to be able to reuse them
