+++ Buform/Platforms/Ios/Components/Switch/SwitchFormCell.cs
@@ -4,11 +4,8 @@ namespace Buform;
 
 [Preserve(AllMembers = true)]
 [Register(nameof(SwitchFormCell))]
-public class SwitchFormCell : FormCell<SwitchFormItem>
+public sealed class SwitchFormCell : SwitchFormCellBase<SwitchFormItem>
 {
-    protected virtual UILabel? Label { get; set; }
-    protected virtual UISwitch? Switch { get; set; }
-
     public SwitchFormCell()
     {
         /* Required constructor */
@@ -18,128 +15,5 @@ public class SwitchFormCell : FormCell<SwitchFormItem>
         : base(handle)
     {
         /* Required constructor */
-    }
-
-    protected override void Initialize()
-    {
-        SelectionStyle = UITableViewCellSelectionStyle.None;
-
-        Label = new UILabel
-        {
-            TranslatesAutoresizingMaskIntoConstraints = false,
-            Font = UIFont.PreferredBody,
-            TextColor = UIColor.Label
-        };
-
-        Switch = new UISwitch();
-        Switch.ValueChanged += OnValueChanged;
-
-        AccessoryView = Switch;
-
-        ContentView.AddSubviews(Label);
-
-        ContentView.AddConstraints(
-            new[]
-            {
-                Label.TopAnchor.ConstraintEqualTo(ContentView.LayoutMarginsGuide.TopAnchor),
-                Label.BottomAnchor.ConstraintEqualTo(ContentView.LayoutMarginsGuide.BottomAnchor),
-                Label.LeadingAnchor.ConstraintEqualTo(ContentView.LayoutMarginsGuide.LeadingAnchor),
-                Label.TrailingAnchor.ConstraintEqualTo(
-                    ContentView.LayoutMarginsGuide.TrailingAnchor
-                )
-            }
-        );
-    }
-
-    private void OnValueChanged(object? sender, EventArgs e)
-    {
-        if (Item == null)
-        {
-            return;
-        }
-
-        Item.Value = Switch?.On ?? false;
-    }
-
-    private void UpdateReadOnlyState()
-    {
-        if (Switch == null)
-        {
-            return;
-        }
-
-        Switch.Enabled = !Item?.IsReadOnly ?? true;
-    }
-
-    private void UpdateLabel()
-    {
-        if (Label == null)
-        {
-            return;
-        }
-
-        Label.Text = Item?.Label;
-    }
-
-    private void UpdateValue(bool isAnimated)
-    {
-        Switch?.SetState(Item?.Value ?? false, isAnimated);
-    }
-
-    private void UpdateValidationErrorMessage()
-    {
-        if (Label == null)
-        {
-            return;
-        }
-
-        Label.TextColor = Item?.ValidationErrorMessage == null ? UIColor.Label : UIColor.SystemRed;
-    }
-
-    protected override void OnItemSet()
-    {
-        UpdateReadOnlyState();
-        UpdateLabel();
-        UpdateValue(false);
-        UpdateValidationErrorMessage();
-    }
-
-    protected override void OnItemPropertyChanged(string? propertyName)
-    {
-        switch (propertyName)
-        {
-            case nameof(Item.IsReadOnly):
-                UpdateReadOnlyState();
-                break;
-            case nameof(Item.Label):
-                UpdateLabel();
-                break;
-            case nameof(Item.Value):
-                UpdateValue(true);
-                break;
-            case nameof(Item.ValidationErrorMessage):
-                UpdateValidationErrorMessage();
-                break;
-        }
-    }
-
-    protected override void Dispose(bool disposing)
-    {
-        if (disposing)
-        {
-            var @switch = Switch;
-            if (@switch != null)
-            {
-                @switch.ValueChanged -= OnValueChanged;
-            }
-
-            Switch?.Dispose();
-            Switch = null;
-
-            Label?.Dispose();
-            Label = null;
-        }
-
-        base.Dispose(disposing);
     }
 }
